Move getPeriodsData to Project model
